* теперь вычисления производятся в long double * Версия 1.0.1.77 релиз :) * список поддерживаемых констант 	"pi" Pi, Archimedes' constant or Ludolph's number 	"c" Speed of light in vacuum (m·s-1) 	"G" Newtonian constant of gravitation (m3·kg−1·s−2), this constant valid on Earths only ;) 	"J" Constants of Gauss fild 	"phi" Golden ratio 	"h" Planck constant (J·s) 	"atm" Standard atmosphere (Pa), this constant valid on Earths only ;) 	"L" Avogadro's number (mol−1) 	"R" Gas constant (J·K−1·mol−1) * список поддерживаемых функций 	"sin" sinus 	"cos" cosinus 	"tan" tangens 	"acos" arccosinus 	"asin" arcsinus 	"atan" arctangens 	"exp" exponent function, on x=1 return value is e 	"log" natural logarithm 	"log10" base-10 logarithm 	"pow" power * поддерживаются операции + - * / и операция степени ^ (эквивалентна функции pow) * поддерживаются выражения со скобками * поддерживаются числа, содержащие степень, например 1.4e-3
git-svn-id: https://studio2000.xyz:8443/svn/TPLTM/Lab1/Calc@77 bea4479c-5f86-c742-a837-c53f34f90b20

+++ Calc/RPN.cpp
@@ -50,10 +50,10 @@ int_8 GetPriority(char p_sym)
 	}
 }
 //---------------------------------------------------------------------------
-void CalculateOnLineExpression(stack<char>& c_operations, stack<double>& c_operands)
+void CalculateOnLineExpression(stack<char>& c_operations, stack<long double>& c_operands)
 {
 	bool set_a = false, set_b = false;
-	double a, b;
+	long double a, b;
 	if(!c_operands.empty())
 	{
 		set_b = true;
@@ -123,7 +123,7 @@ void CalculateLineExpression(string p_input_str, string& p_output_str)
 {
 	int_8 l_current_prioritet = priority_error;
 	stack <char> c_operations;
-	stack <double> c_operands;
+	stack <long double> c_operands;
 	bool l_negative_number;
 	string::size_type l_count = 0;
 	for(; p_input_str.size(); l_count++)
@@ -235,7 +235,7 @@ void CalculateLineExpression(string p_input_str, string& p_output_str)
 			if(!l_negative_number || (l_negative_number && l_count_of_num > 1))
 			{
 				c_operands.push(atof(l_operand_string.c_str()));
-			//	if(c_operands.top() >= std::numeric_limits<double>::max() || c_operands.top() <= std::numeric_limits<double>::min()) // TODO add normal terms ;)
+			//	if(c_operands.top() >= std::numeric_limits<long double>::max() || c_operands.top() <= std::numeric_limits<long double>::min()) // TODO add normal terms ;)
 				if(c_operands.top() > 999999999999999 || c_operands.top() < -999999999999999) // TODO add normal terms ;)
 				{
 					AddWarning(0);
@@ -276,10 +276,10 @@ void CalculateRPN(string& p_to_process_str)
 	 3)Если входной набор символов обработан не полностью, перейти к шагу 1.
 	После полной обработки входного набора символов результат вычисления выражения лежит на вершине стека.
 	*/
-	stack <double> c_operands;
+	stack <long double> c_operands;
 	string input = p_to_process_str;
 	p_to_process_str = "";
-	double a, b;
+	long double a, b;
 	int_8 l_current_priority;
 	
 	for(string::size_type l_count = 0; l_count < input.size(); l_count++)
@@ -389,7 +389,7 @@ void CalculateRPN(string& p_to_process_str)
 									}
 								}
 								
-								double temp_operand = atof(temp.c_str());
+								long double temp_operand = atof(temp.c_str());
 								a = b;
 								b = temp_operand;*/
 							/* 
@@ -439,7 +439,7 @@ void CalculateRPN(string& p_to_process_str)
 	if(!c_operands.empty())
 	{
 		char l_char_buf[100];
-		sprintf_s(l_char_buf, " %.25lf ", c_operands.top());
+		sprintf_s(l_char_buf, " %.50lf ", c_operands.top());
 		p_to_process_str = l_char_buf;
 	}
 }
